feat(extension): extract path trie to dedicated utility module
